Segment level metrics (#835)
Implementing m-bleu, m-chrf3, m-chrf3+, and m-chrf3++ from the paper https://arxiv.org/pdf/2407.12832.

These metrics are implemented at the verse-level, rather than the sentence-level. Currently, they are included in the list of default scoring metrics. We can consider changing this if we don't want so many default metrics (or if these won't be generally useful).